fixed servlets, added auth check

+++ src/main/java/com/borshcheva/shop/web/servlets/DeleteRequestsServlet.java
@@ -1,6 +1,7 @@
 package com.borshcheva.shop.web.servlets;
 
 import com.borshcheva.shop.service.ProductService;
+import com.borshcheva.shop.service.UserAuthService;
 
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -9,16 +10,22 @@ import java.io.IOException;
 
 public class DeleteRequestsServlet extends HttpServlet {
     private final ProductService productService;
+    private final UserAuthService userAuthService;
 
-    public DeleteRequestsServlet(ProductService productService) {
+    public DeleteRequestsServlet(ProductService productService, UserAuthService userAuthService) {
         this.productService = productService;
+        this.userAuthService = userAuthService;
     }
 
     @Override
     public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        int id = Integer.parseInt(request.getParameter("id"));
-        this.productService.deleteProduct(id);
-        response.sendRedirect("/products");
+        if (userAuthService.isAuth(request)) {
+            int id = Integer.parseInt(request.getParameter("id"));
+            this.productService.deleteProduct(id);
+            response.sendRedirect("/products");
+        } else {
+            response.sendRedirect("/login");
+        }
     }
 
 }
